refactor: rename Dto field to align with ContractDefinition
