+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/model/ContractDefinitionRequestDto.java
@@ -37,7 +37,7 @@ public class ContractDefinitionRequestDto extends BaseDto {
     public static final String CONTRACT_DEFINITION_TYPE = EDC_NAMESPACE + "ContractDefinition";
     public static final String CONTRACT_DEFINITION_ACCESSPOLICY_ID = EDC_NAMESPACE + "accessPolicyId";
     public static final String CONTRACT_DEFINITION_CONTRACTPOLICY_ID = EDC_NAMESPACE + "contractPolicyId";
-    public static final String CONTRACT_DEFINITION_CRITERIA = EDC_NAMESPACE + "criteria";
+    public static final String CONTRACT_DEFINITION_SELECTOR_EXPRESSION = EDC_NAMESPACE + "selectorExpression";
     /**
      * Default validity is set to one year.
      */
@@ -46,8 +46,8 @@ public class ContractDefinitionRequestDto extends BaseDto {
     @NotNull(message = "contractPolicyId cannot be null")
     protected String contractPolicyId;
     @Valid
-    @NotNull(message = "criteria cannot be null")
-    protected List<CriterionDto> criteria = new ArrayList<>();
+    @NotNull(message = "selectorExpression cannot be null")
+    protected List<CriterionDto> selectorExpression = new ArrayList<>();
 
     //this cannot be non-null, because that would break backwards compatibility with the old API
     @JsonProperty(value = ID)
@@ -64,8 +64,8 @@ public class ContractDefinitionRequestDto extends BaseDto {
         return contractPolicyId;
     }
 
-    public List<CriterionDto> getCriteria() {
-        return criteria;
+    public List<CriterionDto> getSelectorExpression() {
+        return selectorExpression;
     }
 
     @AssertTrue(message = "id must be either be null or not blank, and it cannot contain the ':' character")
@@ -105,8 +105,8 @@ public class ContractDefinitionRequestDto extends BaseDto {
             return self();
         }
 
-        public Builder criteria(List<CriterionDto> criteria) {
-            dto.criteria = criteria;
+        public Builder selectorExpression(List<CriterionDto> criteria) {
+            dto.selectorExpression = criteria;
             return self();
         }
