+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/transform/ContractDefinitionToContractDefinitionResponseDtoTransformer.java
@@ -44,7 +44,7 @@ public class ContractDefinitionToContractDefinitionResponseDtoTransformer implem
                 .accessPolicyId(object.getAccessPolicyId())
                 .createdAt(object.getCreatedAt())
                 .contractPolicyId(object.getContractPolicyId())
-                .criteria(criteria)
+                .selectorExpression(criteria)
                 .build();
     }
 }
