+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/transform/ContractDefinitionUpdateDtoWrapperToContractDefinitionTransformer.java
@@ -39,7 +39,7 @@ public class ContractDefinitionUpdateDtoWrapperToContractDefinitionTransformer i
 
     @Override
     public @Nullable ContractDefinition transform(@NotNull ContractDefinitionUpdateDtoWrapper object, @NotNull TransformerContext context) {
-        var criteria = object.getContractDefinition().getCriteria().stream().map(it -> context.transform(it, Criterion.class)).collect(Collectors.toList());
+        var criteria = object.getContractDefinition().getSelectorExpression().stream().map(it -> context.transform(it, Criterion.class)).collect(Collectors.toList());
         var selectorExpression = AssetSelectorExpression.Builder.newInstance().criteria(criteria).build();
         return ContractDefinition.Builder.newInstance()
                 .id(object.getId())
