+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/transform/JsonObjectFromContractDefinitionResponseDtoTransformer.java
@@ -25,7 +25,7 @@ import org.jetbrains.annotations.Nullable;
 
 import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_ACCESSPOLICY_ID;
 import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_CONTRACTPOLICY_ID;
-import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_CRITERIA;
+import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_SELECTOR_EXPRESSION;
 import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_TYPE;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.ID;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
@@ -47,12 +47,12 @@ public class JsonObjectFromContractDefinitionResponseDtoTransformer extends Abst
                 .add(CONTRACT_DEFINITION_CONTRACTPOLICY_ID, dto.getContractPolicyId());
 
         // todo: does the criterionDto need to be json-ld-ified?
-        var criteria = dto.getCriteria().stream()
+        var criteria = dto.getSelectorExpression().stream()
                 .map(criterion -> context.transform(criterion, JsonObject.class))
                 .collect(jsonFactory::createArrayBuilder, JsonArrayBuilder::add, JsonArrayBuilder::add)
                 .build();
 
-        builder.add(CONTRACT_DEFINITION_CRITERIA, criteria);
+        builder.add(CONTRACT_DEFINITION_SELECTOR_EXPRESSION, criteria);
         return builder.build();
     }
 }
