+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/transform/JsonObjectToContractDefinitionRequestDtoTransformer.java
@@ -27,7 +27,7 @@ import java.util.ArrayList;
 
 import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_ACCESSPOLICY_ID;
 import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_CONTRACTPOLICY_ID;
-import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_CRITERIA;
+import static org.eclipse.edc.connector.api.management.contractdefinition.model.ContractDefinitionRequestDto.CONTRACT_DEFINITION_SELECTOR_EXPRESSION;
 
 public class JsonObjectToContractDefinitionRequestDtoTransformer extends AbstractJsonLdTransformer<JsonObject, ContractDefinitionRequestDto> {
 
@@ -49,10 +49,10 @@ public class JsonObjectToContractDefinitionRequestDtoTransformer extends Abstrac
             transformString(value, builder::accessPolicyId, context);
         } else if (CONTRACT_DEFINITION_CONTRACTPOLICY_ID.equals(key)) {
             transformString(value, builder::contractPolicyId, context);
-        } else if (CONTRACT_DEFINITION_CRITERIA.equals(key)) {
+        } else if (CONTRACT_DEFINITION_SELECTOR_EXPRESSION.equals(key)) {
             var list = new ArrayList<CriterionDto>();
             transformArrayOrObject(value, CriterionDto.class, list::add, context);
-            builder.criteria(list);
+            builder.selectorExpression(list);
         }
     }
 }
