+++ extensions/control-plane/api/management-api/contract-definition-api/src/test/java/org/eclipse/edc/connector/api/management/contractdefinition/model/ContractDefinitionRequestDtoTest.java
@@ -36,7 +36,7 @@ class ContractDefinitionRequestDtoTest {
                 .contractPolicyId("test-contract-policyid")
                 .accessPolicyId("test-access-policyid")
                 .id("test-id")
-                .criteria(List.of(criterion))
+                .selectorExpression(List.of(criterion))
                 .build();
 
         var str = objectMapper.writeValueAsString(dto);
