+++ extensions/control-plane/api/management-api/contract-definition-api/src/test/java/org/eclipse/edc/connector/api/management/contractdefinition/model/ContractDefinitionRequestDtoValidationTest.java
@@ -49,7 +49,7 @@ class ContractDefinitionRequestDtoValidationTest {
                 .id(id)
                 .accessPolicyId(accessPolicyId)
                 .contractPolicyId(contractPolicyId)
-                .criteria(criteria)
+                .selectorExpression(criteria)
                 .build();
 
         var result = validator.validate(dto);
@@ -64,7 +64,7 @@ class ContractDefinitionRequestDtoValidationTest {
                 .id(id)
                 .accessPolicyId(accessPolicyId)
                 .contractPolicyId(contractPolicyId)
-                .criteria(criteria)
+                .selectorExpression(criteria)
                 .build();
 
         var result = validator.validate(dto);
