+++ extensions/control-plane/api/management-api/contract-definition-api/src/test/java/org/eclipse/edc/connector/api/management/contractdefinition/transform/ContractDefinitionToContractDefinitionResponseDtoTransformerTest.java
@@ -59,7 +59,7 @@ class ContractDefinitionToContractDefinitionResponseDtoTransformerTest {
         assertThat(dto.getAccessPolicyId()).isEqualTo(contractDefinition.getAccessPolicyId());
         assertThat(dto.getContractPolicyId()).isEqualTo(contractDefinition.getContractPolicyId());
         assertThat(dto.getCreatedAt()).isNotZero();
-        assertThat(dto.getCriteria()).containsOnly(criterionDto);
+        assertThat(dto.getSelectorExpression()).containsOnly(criterionDto);
         verify(context).transform(isA(Criterion.class), eq(CriterionDto.class));
     }
