+++ resources/openapi/yaml/management-api/policy-definition-api.yaml
@@ -1,251 +1,5 @@
 openapi: 3.0.1
 paths:
-  /policydefinitions:
-    get:
-      deprecated: true
-      description: Returns all policy definitions according to a query
-      operationId: getAllPolicies
-      parameters:
-      - in: query
-        name: offset
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: limit
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: sort
-        schema:
-          type: string
-          enum:
-          - ASC
-          - DESC
-          example: null
-      - in: query
-        name: sortField
-        schema:
-          type: string
-          example: null
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/PolicyDefinitionResponseDto'
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request was malformed
-      tags:
-      - Policy
-    post:
-      deprecated: true
-      description: Creates a new policy definition
-      operationId: createPolicy
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/PolicyDefinitionRequestDto'
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/IdResponseDto'
-          description: policy definition was created successfully. Returns the Policy
-            Definition Id and created timestamp
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request body was malformed
-        "409":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: "Could not create policy definition, because a contract definition\
-            \ with that ID already exists"
-      tags:
-      - Policy
-  /policydefinitions/request:
-    post:
-      deprecated: true
-      description: Returns all policy definitions according to a query
-      operationId: queryAllPolicies
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/QuerySpecDto'
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/PolicyDefinitionResponseDto'
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request was malformed
-      tags:
-      - Policy
-  /policydefinitions/{id}:
-    delete:
-      deprecated: true
-      description: "Removes a policy definition with the given ID if possible. Deleting\
-        \ a policy definition is only possible if that policy definition is not yet\
-        \ referenced by a contract definition, in which case an error is returned.\
-        \ DANGER ZONE: Note that deleting policy definitions can have unexpected results,\
-        \ do this at your own risk!"
-      operationId: deletePolicy
-      parameters:
-      - in: path
-        name: id
-        required: true
-        schema:
-          type: string
-          example: null
-      responses:
-        "200":
-          description: Policy definition was deleted successfully
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: "Request was malformed, e.g. id was null"
-        "404":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: An policy definition with the given ID does not exist
-        "409":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: "The policy definition cannot be deleted, because it is referenced\
-            \ by a contract definition"
-      tags:
-      - Policy
-    get:
-      deprecated: true
-      description: Gets a policy definition with the given ID
-      operationId: getPolicy
-      parameters:
-      - in: path
-        name: id
-        required: true
-        schema:
-          type: string
-          example: null
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/PolicyDefinitionResponseDto'
-          description: The  policy definition
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: "Request was malformed, e.g. id was null"
-        "404":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: An  policy definition with the given ID does not exist
-      tags:
-      - Policy
-  /policydefinitions/{policyId}:
-    put:
-      deprecated: true
-      description: "Updates an existing Policy, If the Policy is not found, an error\
-        \ is reported"
-      operationId: updatePolicy
-      parameters:
-      - in: path
-        name: policyId
-        required: true
-        schema:
-          type: string
-          example: null
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/PolicyDefinitionUpdateDto'
-      responses:
-        "200":
-          description: policy definition was updated successfully. Returns the Policy
-            Definition Id and updated timestamp
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request body was malformed
-        "404":
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ApiErrorDetail'
-          description: "policy definition could not be updated, because it does not\
-            \ exists"
-      tags:
-      - Policy
   /v2/policydefinitions:
     post:
       description: Creates a new policy definition
