+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/Participant.java
@@ -130,7 +130,7 @@ public class Participant {
                 .add(TYPE, EDC_NAMESPACE + "ContractDefinition")
                 .add(EDC_NAMESPACE + "accessPolicyId", accessPolicyId)
                 .add(EDC_NAMESPACE + "contractPolicyId", contractPolicyId)
-                .add(EDC_NAMESPACE + "criteria", Json.createArrayBuilder()
+                .add(EDC_NAMESPACE + "selectorExpression", Json.createArrayBuilder()
                         .add(createObjectBuilder()
                                 .add(TYPE, "CriterionDto")
                                 .add(EDC_NAMESPACE + "operandLeft", EDC_NAMESPACE + "id")
